requirements_dev.txt
pytz
aiohttp
aiomqtt
pytest
pytest-asyncio
aioresponses
python-dotenv
rich
